feat: 파일 분리, pickUpTable 튜플형식으로 구현
